Split the command Check into a pure core and a spawn shell
command.ts held the whole redproof/command subpath in one file: option
validation, exit-code policy, result mapping, group aggregation, the
spawn loop, and a private copy of mapLimit. Every rule about exit codes
was only reachable by spawning a real process.

Make command a bounded context. command/core holds the decisions as
plain functions: exitPolicy and classifyExit, option validation and
descriptions, isInsideRoot, commandResult, and aggregateResults, with
timestamps passed in. command/shell/spawn.ts holds the child process
and its timers. command/shell/check.ts wires the two and reads the
clock. The subpath export now points at dist/command/index.js.

mapLimit had two copies, in command.ts and run/shell/worker-process.ts.
One copy lives in support/map-limit.ts.
