+++ gates/support/effects-model.ts
@@ -27,7 +27,8 @@ const APPROVED_EFFECT_BOUNDARIES = new Set([
   'packages/dependency-cruiser/src/index.ts',
   'packages/eslint/src/index.ts',
   'packages/redproof/src/cli.ts',
-  'packages/redproof/src/command.ts',
+  'packages/redproof/src/command/shell/check.ts',
+  'packages/redproof/src/command/shell/spawn.ts',
   'packages/redproof/src/composition/inspect.ts',
   'packages/redproof/src/composition/mutation/context.ts',
   'packages/redproof/src/composition/mutation/filesystem.ts',
