+++ packages/redproof/src/command/core/options.ts
@@ -1,0 +1,77 @@
+import { basename } from 'node:path';
+import type { Rule, RuleRef } from '../../domain/rule.ts';
+import type { CommandExitCodes } from './exit-codes.ts';
+
+export const DEFAULT_MAX_OUTPUT_BYTES = 10 * 1024 * 1024;
+export const DEFAULT_MAX_AT_ONCE = 4;
+
+export type CommandCheckOptions<R extends RuleRef> = {
+  readonly rule: Rule<R>;
+  readonly command: string;
+  readonly args?: readonly string[];
+  readonly label?: string;
+  readonly description?: string;
+  /** Relative to the Gate root and confined inside it. Defaults to the root. */
+  readonly cwd?: string;
+  readonly env?: Readonly<Record<string, string | undefined>>;
+  readonly timeoutMs?: number;
+  /** Combined stdout and stderr capture limit. Defaults to 10 MiB. */
+  readonly maxOutputBytes?: number;
+  readonly exitCodes?: CommandExitCodes;
+};
+
+export type CommandGroupExecution =
+  | {
+      readonly mode?: 'sequential';
+      readonly maxAtOnce?: never;
+    }
+  | {
+      readonly mode: 'parallel';
+      /** Maximum number of child processes running together. Defaults to 4. */
+      readonly maxAtOnce?: number;
+    };
+
+export type CommandGroupOptions<R extends RuleRef> = {
+  readonly entries: readonly [CommandCheckOptions<R>, ...CommandCheckOptions<R>[]];
+  readonly label?: string;
+  readonly description?: string;
+} & CommandGroupExecution;
+
+function validatePositiveInteger(name: string, value: number | undefined): void {
+  if (value !== undefined && (!Number.isSafeInteger(value) || value <= 0)) {
+    throw new Error(`${name} must be a positive integer.`);
+  }
+}
+
+export function validateCommandOptions(options: CommandCheckOptions<RuleRef>): void {
+  if (!options.command.trim()) throw new Error('command must not be empty.');
+  validatePositiveInteger('timeoutMs', options.timeoutMs);
+  validatePositiveInteger('maxOutputBytes', options.maxOutputBytes);
+}
+
+export function validateGroupOptions(options: CommandGroupOptions<RuleRef>): void {
+  if (options.entries.length === 0) throw new Error('commands requires at least one entry.');
+  if (options.mode !== 'parallel' && options.maxAtOnce !== undefined) {
+    throw new Error('maxAtOnce is only available in parallel mode.');
+  }
+  if (options.mode === 'parallel') validatePositiveInteger('maxAtOnce', options.maxAtOnce);
+}
+
+export function labelOf(options: CommandCheckOptions<RuleRef>): string {
+  return options.label ?? options.command;
+}
+
+export function commandDescription(options: CommandCheckOptions<RuleRef>): string {
+  const invocation = [basename(options.command), ...(options.args ?? [])].join(' ');
+  return options.description ?? `run ${invocation}`;
+}
+
+export function groupDescription(options: CommandGroupOptions<RuleRef>): string {
+  return options.description
+    ?? `run ${options.entries.length} commands: `
+      + options.entries.map(entry => entry.label ?? basename(entry.command)).join(', ');
+}
+
+export function groupSource(options: CommandGroupOptions<RuleRef>): string {
+  return options.label ?? options.entries.map(labelOf).join(', ');
+}
