+++ packages/redproof/src/command/index.ts
@@ -1,0 +1,3 @@
+export type { CommandExitCodes } from './core/exit-codes.ts';
+export type { CommandCheckOptions, CommandGroupExecution, CommandGroupOptions } from './core/options.ts';
+export { command, commands } from './shell/check.ts';
