+++ packages/redproof/src/run/shell/worker-process.ts
@@ -2,26 +2,6 @@ import { fork } from 'node:child_process';
 import { fileURLToPath } from 'node:url';
 import type { GateWorkerJob, GateWorkerResult } from '../core/protocol.ts';
 
-export async function mapLimit<T, R>(
-  items: readonly T[],
-  limit: number,
-  run: (item: T, index: number) => Promise<R>,
-): Promise<R[]> {
-  const results = new Array<R>(items.length);
-  let cursor = 0;
-
-  async function worker(): Promise<void> {
-    while (true) {
-      const index = cursor++;
-      if (index >= items.length) return;
-      results[index] = await run(items[index]!, index);
-    }
-  }
-
-  await Promise.all(Array.from({ length: Math.min(limit, items.length) }, () => worker()));
-  return results;
-}
-
 const workerExtension = import.meta.url.endsWith('.ts') ? 'ts' : 'js';
 const gateWorkerFile = fileURLToPath(
   new URL(`./gate-worker.${workerExtension}`, import.meta.url),
